# Native Backend Optimization: Moved *_base offset math outside the inner loop.