speed-ups for every dtype except int64 and complex64 (int64 path rolled back for now).

ops_binary.ml is ~8 k LOC. Before copying this change to every op, should we switch to a functor/code-gen approach to reduce duplication?     Only  **add** is optimized in this PR.

Included the benchmark file I used to compare the old and new add kernels.

# Benchmark Comparison (Before vs After Optimization)

| Benchmark             |   Before (μs) |   After (μs) |   Δ (μs) |   Improvement (%) |   Speedup (x) |
|:----------------------|--------------:|-------------:|---------:|------------------:|--------------:|
| Add 500x500 float32   |        381.76 |       162.1  |   219.66 |             57.54 |          2.36 |
| Add 500x500 uint16    |        369.89 |       158.59 |   211.3  |             57.13 |          2.33 |
| Add 500x500 int16     |        359.99 |       156.69 |   203.3  |             56.47 |          2.3  |
| Add 500x500 int32     |        374.67 |       163.37 |   211.3  |             56.4  |          2.29 |
| Add 500x500 int8      |        366.41 |       162.11 |   204.3  |             55.76 |          2.26 |
| Add 500x500 uint8     |        364.89 |       169.32 |   195.57 |             53.6  |          2.16 |
| Add 500x500 float64   |        416.97 |       220.41 |   196.56 |             47.14 |          1.89 |
| Add 200x200 uint8     |         83.63 |        47    |    36.63 |             43.8  |          1.78 |
| Add 200x200 int8      |         82.73 |        46.85 |    35.88 |             43.37 |          1.77 |
| Add 200x200 uint16    |        110.51 |        69.75 |    40.76 |             36.88 |          1.58 |
| Add 100x100 uint16    |         41.78 |        27.56 |    14.22 |             34.04 |          1.52 |
| Add 100x100 int16     |         42.08 |        28.08 |    14    |             33.27 |          1.5  |
| Add 100x100 float32   |         42.28 |        28.65 |    13.63 |             32.24 |          1.48 |
| Add 200x200 int16     |        104.97 |        71.97 |    33    |             31.44 |          1.46 |
| Add 100x100 uint8     |         40.03 |        27.68 |    12.35 |             30.85 |          1.45 |
| Add 100x100 int32     |         43.46 |        30.09 |    13.37 |             30.76 |          1.44 |
| Add 200x200 float32   |        127.43 |        89.22 |    38.21 |             29.99 |          1.43 |
| Add 200x200 float64   |        112.35 |        80.88 |    31.47 |             28.01 |          1.39 |
| Add 200x200 int32     |        117.19 |        85.62 |    31.57 |             26.94 |          1.37 |
| Add 100x100 int8      |         39.09 |        30.19 |     8.9  |             22.77 |          1.29 |
| Add 500x500 float16   |        462.2  |       357.56 |   104.64 |             22.64 |          1.29 |
| Add 50x50 int32       |         25.64 |        20.76 |     4.88 |             19.03 |          1.24 |
| Add 200x200 complex32 |        196.77 |       160.92 |    35.85 |             18.22 |          1.22 |
| Add 50x50 int16       |         26.17 |        21.65 |     4.52 |             17.27 |          1.21 |
| Add 50x50 int64       |         25.72 |        21.35 |     4.37 |             16.99 |          1.2  |
| Add 200x200 complex64 |        210.96 |       176.92 |    34.04 |             16.14 |          1.19 |
| Add 50x50 float64     |         24.58 |        20.9  |     3.68 |             14.97 |          1.18 |
| Add 50x50 float32     |         24.93 |        21.2  |     3.73 |             14.96 |          1.18 |
| Add 100x100 float64   |         58.45 |        49.9  |     8.55 |             14.63 |          1.17 |
| Add 50x50 complex32   |         30.45 |        26.7  |     3.75 |             12.32 |          1.14 |
| Add 50x50 uint8       |         24.24 |        21.35 |     2.89 |             11.92 |          1.14 |
| Add 100x100 complex32 |         82.06 |        73.84 |     8.22 |             10.02 |          1.11 |
| Add 50x50 int8        |         23.73 |        21.76 |     1.97 |              8.3  |          1.09 |
| Add 500x500 complex32 |        662.82 |       616.11 |    46.71 |              7.05 |          1.08 |
| Add 200x200 float16   |        125.03 |       116.81 |     8.22 |              6.57 |          1.07 |
| Add 50x50 uint16      |         23.28 |        21.82 |     1.46 |              6.27 |          1.07 |
| Add 50x50 complex64   |         30.96 |        29.86 |     1.1  |              3.55 |          1.04 |
| Add 50x50 float16     |         27.39 |        26.74 |     0.65 |              2.37 |          1.02 |
| Add 100x100 float16   |         45.94 |        44.96 |     0.98 |              2.13 |          1.02 |
| Add 100x100 complex64 |        107.31 |       105.42 |     1.89 |              1.76 |          1.02 |
| Add 500x500 complex64 |        891.75 |       914.73 |   -22.98 |             -2.58 |          0.97 |
| Add 500x500 int64     |        415.13 |       621.75 |  -206.62 |            -49.77 |          0.67 |
| Add 100x100 int64     |         70.97 |       119.95 |   -48.98 |            -69.02 |          0.59 |
| Add 200x200 int64     |        127.73 |       262.99 |  -135.26 |           -105.9  |          0.49 |