@tmattio 2025-06-22
Hi @blueavee! Thanks a lot for your work on this.

I can confirm that the **massive** code duplication in the native backend is intentional and necessary, I've seen a 5X performance improvement implementing non-polymorphic functions. I've explored a lot of alternatives, including functors, but all of them came with a huge performance hit. This is slighly annoying as I suspect this is the main culprit for the slow compilation time of Raven, but I guess we'll have to just accept it.

C.f. https://ocaml.org/manual/5.3/flambda.html that hints at the fact that specialization is necessary for performance, but given the size of our function, the compiler seems to decide not to specialize our functions.

Side note I wonder if parametric functions could be useful here (https://github.com/ocaml/ocaml/pull/13806). cc @voodoos - sorry for the random cc, feel free to ignore if this doesn't seem relevant.